# internal/api/router/openapi.yaml
openapi: "3.1.0"
info:
  title: QPKI - Post-Quantum PKI REST API
  version: "1.0.0"
  description: |
    REST API for Post-Quantum Public Key Infrastructure (PKI).

    Supports post-quantum cryptographic algorithms:
    - ML-DSA (Dilithium) for signatures
    - SLH-DSA (SPHINCS+) for hash-based signatures
    - ML-KEM (Kyber) for key encapsulation
    - Hybrid modes combining classical and PQ algorithms

    All binary data is encoded as Base64 with optional PEM encoding.
  contact:
    name: QPKI
    url: https://github.com/remiblancher/post-quantum-pki
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8443
    description: Local development server
  - url: https://{host}:{port}
    description: Production server
    variables:
      host:
        default: pki.example.com
      port:
        default: "443"

# Security is optional by default (empty array means no auth required)
# Future: Add PQ signature-based auth
security:
  - {}
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/cwt/issue:
    post:
      tags: [CWT]
      summary: Issue a CBOR Web Token
      operationId: cwtIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CWTIssueRequest"
      responses:
        "200":
          description: Issued CWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CWTIssueResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/cwt/verify:
    post:
      tags: [CWT]
      summary: Verify a CBOR Web Token
      operationId: cwtVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CWTVerifyRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CWTVerifyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # TSA endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (optional)
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (future PQ signature)
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
  schemas:
...
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Error code (e.g., CERT_NOT_FOUND)
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
    BinaryData:
      type: object
      required: [data]
      properties:
        data:
          type: string
          description: Base64-encoded data
        encoding:
          type: string
          enum: [base64, pem]
          default: base64
...
    CWTIssueRequest:
      type: object
      required: [credential_id, claims]
      properties:
        credential_id:
          type: string
        passphrase:
          type: string
        claims:
          type: object
          properties:
            iss:
              type: string
            sub:
              type: string
            aud:
              type: string
            exp:
              type: integer
            nbf:
              type: integer
            iat:
              type: integer
            cti:
              type: string
            custom:
              type: object
              additionalProperties: true
    CWTIssueResponse:
      type: object
      properties:
        cwt:
          $ref: "#/components/schemas/BinaryData"
    CWTVerifyRequest:
      type: object
      required: [cwt]
      properties:
        cwt:
          $ref: "#/components/schemas/BinaryData"
        trusted_certs:
          type: array
          items:
            $ref: "#/components/schemas/BinaryData"
    CWTVerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
        claims:
          type: object
        errors:
          type: array
          items:
            type: string

    # TSA schemas
...